+++ CHANGES.md
@@ -22,6 +22,7 @@
   - ➕ `localeCompare` / `locale_compare`
   - ➕ `normalize`
   - ➕ `padEnd` / `pad_end`
+  - ➕ `padStart` / `pad_start`
 
 ## 0.0.1 - 2023-04-28